Question 1: After installing the "palmgerpenguins" in the console, I used the code provided to me in the homework to load the `penguins` dataset in my RMarkdown file.
I used `view(penguins)` in my console to view the entire dataset. I used `penguins` to include the code chunk of the dataset when the file is knit.

I made a list to represent the names of the variables included in the dataset.

I used `r` inline code so the number of rows, the number of columns, and the mean flipper length appears as text when the RMarkdown is knit.

I created a scattplot using ggplot and exported the plot into my project directory using ggsave.